Created a Media page
Created files for the Media page.
